This is the program at its least amount of functionability
the green column of buttons is for choosing the shot you are trying to do.

The grey column is for choosing the shot you actually ended up hitting.

then you hit next hole.
and choose again.

after 18 selections there is a alert to demonstrate that the values are being stored into two seperate arrays.